requirements.txt
blinker==1.6.3
cffi==1.17.1
click==8.1.7
cryptography==45.0.5
dotenv==0.9.9
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
protobuf==4.21.12
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==2.3.7
gunicorn
